+++ latch_cli/snakemake/single_task_snakemake.py
@@ -6,6 +6,7 @@ from textwrap import dedent
 from typing import Dict, Set
 
 import snakemake
+import snakemake.workflow
 from snakemake.parser import (
     INDENT,
     Benchmark,
@@ -36,6 +37,12 @@ if print_compilation == "1":
 data = json.loads(os.environ["LATCH_SNAKEMAKE_DATA"])
 rules = data["rules"]
 outputs = data["outputs"]
+
+non_blob_parameters = data.get("non_blob_parameters", {})
+
+# todo(ayush): do this without overwriting globals
+sw = sys.modules["snakemake.workflow"]
+setattr(sw, "config", non_blob_parameters)
 
 
 def eprint_named_list(xs):